Remove some redundant casts, deprecated methods, and parametise.

+++ ORAC/src/orac/ukirt/inst/SpInstCGS4.java
@@ -547,7 +547,7 @@ public final class SpInstCGS4 extends SpUKIRTInstObsComp
 		{
 			System.out.println( "Error calculating default no. coadds" ) ;
 		}
-		return ( int )coadds ;
+		return coadds ;
 	}
 
 	/**
